requirements.txt
flask
pandas
google-generativeai
python-dotenv
flask-session